# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MM Swagger Api",
        "contact": {
            "email": "alexey@sportysystems.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:1000/api",
            "description": "Dev MM Server"
        },
        {
            "url": "http://localhost:1000/",
            "description": "Local MM Server"
        }
    ],
    "paths": {
...
        "/evaluate": {
            "post": {
                "tags": [
                    "evaluate"
                ],
                "operationId": "evaluate.store",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EvaluationScore"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/evaluate-virtual": {
            "post": {
                "tags": [
                    "evaluate"
                ],
                "operationId": "evaluate.virtual",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EvaluationVirtual"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EvaluationVirtual"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "EvaluationScore": {
                "title": "Evaluation Score",
                "description": "Evaluation Score",
                "properties": {
                    "fsw": {
                        "type": "integer",
                        "format": "number"
                    },
                    "crs": {
                        "type": "integer",
                        "format": "number"
                    },
                    "result": {
                        "type": "string",
                        "format": "string"
                    },
                    "has_high_crs": {
                        "type": "boolean",
                        "format": "boolean"
                    }
                },
                "type": "object"
            },
            "AbroadStudyExperience": {
                "properties": {
                    "type_of_program": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "EvaluationVirtual": {
                "properties": {
                    "abroadStudyExperience": {
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/AbroadStudyExperience"
                        }
                    },
                    "canadianJobOffer": {
                        "$ref": "#/components/schemas/VirtualCanadianJobOffer"
                    },
                    "canadianStudyExperience": {
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/VirtualCanadianStudyExperience"
                        }
                    },
                    "canadianWorkExperience": {
                        "format": "array",
                        "items": {
                            "$ref": "#/components/schemas/VirtualCanadianWorkExperience"
                        }
                    },
                    "entrepreneurship": {
                        "$ref": "#/components/schemas/VirtualEntrepreneurship"
                    },
                    "language": {
                        "$ref": "#/components/schemas/VirtualLanguageEvaluate"
                    },
                    "relatives": {
                        "$ref": "#/components/schemas/VirtualRelativesEvaluate"
                    },
                    "spouse": {
                        "$ref": "#/components/schemas/VirtualSpouseEvaluate"
                    },
                    "work": {
                        "$ref": "#/components/schemas/VirtualWorkEvaluate"
                    }
                },
                "type": "object"
            },
            "VirtualCanadianJobOffer": {
                "properties": {
                    "occupation": {
                        "type": "string"
                    },
                    "province": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "VirtualCanadianStudyExperience": {
                "properties": {
                    "completed": {
                        "type": "boolean"
                    },
                    "completion_date": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "mb_field_in_steam": {
                        "type": "string"
                    },
                    "mb_steam_internship": {
                        "type": "string"
                    },
                    "province": {
                        "type": "string"
                    },
                    "type_of_program": {
                        "type": "string"
                    },
                    "institution": {
                        "type": "string"
                    },
                    "program_name": {
                        "type": "string"
                    },
                    "field_of_study": {
                        "type": "string"
                    },
                    "mb_bridging_program": {
                        "type": "string"
                    },
                    "resided_16_months_in_atlantic_province": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "VirtualCanadianWorkExperience": {
                "properties": {
                    "occupation": {
                        "type": "boolean"
                    },
                    "province": {
                        "type": "string"
                    },
                    "when": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "VirtualEntrepreneurship": {
                "properties": {
                    "business_owner_experience": {
                        "type": "boolean"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "net_worth": {
                        "type": "integer"
                    },
                    "senior_manager_experience": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "VirtualLanguageEvaluate": {
                "properties": {
                    "english": {
                        "$ref": "#/components/schemas/VirtualLanguageEvaluateItem"
                    },
                    "french": {
                        "$ref": "#/components/schemas/VirtualLanguageEvaluateItem"
                    }
                },
                "type": "object"
            },
            "VirtualLanguageEvaluateItem": {
                "properties": {
                    "clb": {
                        "type": "integer"
                    },
                    "language_test": {
                        "type": "boolean"
                    },
                    "listening": {
                        "type": "integer"
                    },
                    "reading": {
                        "type": "integer"
                    },
                    "speaking": {
                        "type": "integer"
                    },
                    "writing": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "VirtualRelativesEvaluate": {
                "properties": {
                    "canadian_status": {
                        "type": "string"
                    },
                    "province": {
                        "type": "string"
                    },
                    "relationship": {
                        "type": "string"
                    },
                    "residency_duration": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "VirtualSpouseEvaluate": {
                "properties": {
                    "english": {
                        "type": "number"
                    },
                    "french": {
                        "type": "number"
                    },
                    "inside": {
                        "$ref": "#/components/schemas/VirtualSpouseEvaluateInside"
                    },
                    "outside": {
                        "$ref": "#/components/schemas/VirtualSpouseEvaluateOutside"
                    }
                },
                "type": "object"
            },
            "VirtualSpouseEvaluateInside": {
                "properties": {
                    "highschool_completed": {
                        "type": "boolean"
                    },
                    "type_of_program": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    },
                    "skilled": {
                        "type": "number",
                        "format": "double"
                    },
                    "unskilled": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
            "VirtualSpouseEvaluateOutside": {
                "properties": {
                    "highschool_completed": {
                        "type": "boolean"
                    },
                    "province": {
                        "type": "string"
                    },
                    "type_of_program": {
                        "type": "string"
                    },
                    "duration": {
                        "type": "string"
                    },
                    "skilled": {
                        "type": "number",
                        "format": "double"
                    },
                    "unskilled": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
            "VirtualWorkEvaluate": {
                "properties": {
                    "skilled": {
                        "type": "string"
                    },
                    "unskilled": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}